I made the necessary changes for me to be able to run workflow tests on LUMI and for this branch to pass the tests. The main reason for failures was the new run_dir structure. get_run_dir_count needed to be modified with the new regex.
