+++ workflow_tests/automated_tests_no_HPC/test_sequential_workflow.py
@@ -3,7 +3,7 @@ Basic tests for the sequential workflow.
 """
 
 import pytest
-from workflow_tests.utils.test_utils import *
+from workflow_tests.utils.test_utils import get_run_dir_count, read_summary_file
 from enchanted_surrogates.supervisor.supervisor import Supervisor
 
 def test_sequential_workflow(tmp_path, run_config):
